--- src/main/java/com/atlassian/bitbucket/jenkins/internal/applink/oauth/provider/StoreException.java
@@ -1,19 +1,0 @@
-package com.atlassian.bitbucket.jenkins.internal.applink.oauth.provider;
-
-/**
- * Thrown if there is a problem storing or loading values.
- */
-public class StoreException extends RuntimeException {
-
-    public StoreException(String message, Throwable cause) {
-        super(message, cause);
-    }
-
-    public StoreException(String message) {
-        super(message);
-    }
-
-    public StoreException(Throwable cause) {
-        super(cause);
-    }
-}
