+++ src/main/java/com/atlassian/bitbucket/jenkins/internal/applink/oauth/serviceprovider/exception/StoreException.java
@@ -1,0 +1,11 @@
+package com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.exception;
+
+/**
+ * Thrown if there is a problem storing or loading values.
+ */
+public class StoreException extends RuntimeException {
+
+    public StoreException(String message) {
+        super(message);
+    }
+}
