+++ src/main/java/com/atlassian/bitbucket/jenkins/internal/applink/oauth/serviceprovider/rest/AccessTokenRestEndpoint.java
@@ -1,10 +1,10 @@
-package com.atlassian.bitbucket.jenkins.internal.applink.oauth.rest;
+package com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.rest;
 
 import com.atlassian.bitbucket.jenkins.internal.applink.oauth.adaptor.OAuthConverter;
-import com.atlassian.bitbucket.jenkins.internal.applink.oauth.provider.InvalidTokenException;
-import com.atlassian.bitbucket.jenkins.internal.applink.oauth.provider.ServiceProviderToken;
-import com.atlassian.bitbucket.jenkins.internal.applink.oauth.provider.ServiceProviderTokenStore;
-import com.atlassian.bitbucket.jenkins.internal.applink.oauth.provider.TokenFactory;
+import com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.exception.InvalidTokenException;
+import com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.token.ServiceProviderToken;
+import com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.token.ServiceProviderTokenStore;
+import com.atlassian.bitbucket.jenkins.internal.applink.oauth.serviceprovider.token.TokenFactory;
 import net.oauth.OAuthMessage;
 import net.oauth.OAuthProblemException;
 import net.oauth.OAuthValidator;
@@ -72,7 +72,7 @@ public class AccessTokenRestEndpoint {
             }
 
             try {
-                oAuthValidator.validateMessage(requestMessage, OAuthConverter.asOAuthAccessor(token));
+                oAuthValidator.validateMessage(requestMessage, OAuthConverter.createOAuthAccessor(token));
             } catch (OAuthProblemException ope) {
                 logOAuthProblem(requestMessage, ope, LOGGER);
                 throw ope;
